Follow XDG base directory specification for swiftly's home and bin directories on Linux (#26)

+++ Sources/LinuxPlatform/Linux.swift
@@ -5,22 +5,16 @@ import SwiftlyCore
 /// This implementation can be reused for any supported Linux platform.
 /// TODO: replace dummy implementations
 public struct Linux: Platform {
-    private let platform: Config.PlatformDefinition
+    public init() {}
 
-    public init(platform: Config.PlatformDefinition) {
-        self.platform = platform
-    }
-
-    public var name: String {
-        self.platform.name
-    }
-
-    public var nameFull: String {
-        self.platform.nameFull
-    }
-
-    public var namePretty: String {
-        self.platform.namePretty
+    public var appDataDirectory: URL {
+        if let dir = ProcessInfo.processInfo.environment["XDG_DATA_HOME"] {
+            return URL(fileURLWithPath: dir)
+        } else {
+            return FileManager.default.homeDirectoryForCurrentUser
+                .appendingPathComponent(".local", isDirectory: true)
+                .appendingPathComponent("share", isDirectory: true)
+        }
     }
 
     public var toolchainFileExtension: String {
@@ -36,13 +30,12 @@ public struct Linux: Platform {
             throw Error(message: "\(tmpFile) doesn't exist")
         }
 
-        let toolchainsDir = SwiftlyCore.homeDir.appendingPathComponent("toolchains")
-        if !toolchainsDir.fileExists() {
-            try FileManager.default.createDirectory(at: toolchainsDir, withIntermediateDirectories: false)
+        if !self.swiftlyToolchainsDir.fileExists() {
+            try FileManager.default.createDirectory(at: self.swiftlyToolchainsDir, withIntermediateDirectories: false)
         }
 
         SwiftlyCore.print("Extracting toolchain...")
-        let toolchainDir = toolchainsDir.appendingPathComponent(version.name)
+        let toolchainDir = self.swiftlyToolchainsDir.appendingPathComponent(version.name)
 
         if toolchainDir.fileExists() {
             try FileManager.default.removeItem(at: toolchainDir)
@@ -52,40 +45,85 @@ public struct Linux: Platform {
             // drop swift-a.b.c-RELEASE etc name from the extracted files.
             let relativePath = name.drop { c in c != "/" }.dropFirst()
 
-            // prepend ~/.swiftly/toolchains/<toolchain> to each file name
+            // prepend /path/to/swiftlyHomeDir/toolchains/<toolchain> to each file name
             return toolchainDir.appendingPathComponent(String(relativePath))
         }
     }
 
     public func uninstall(_ toolchain: ToolchainVersion) throws {
-        let toolchainDir = SwiftlyCore.toolchainsDir.appendingPathComponent(toolchain.name)
+        let toolchainDir = self.swiftlyToolchainsDir.appendingPathComponent(toolchain.name)
         try FileManager.default.removeItem(at: toolchainDir)
     }
 
-    public func use(_ toolchain: ToolchainVersion) throws {
-        let toolchainBinURL = SwiftlyCore.toolchainsDir
+    public func use(_ toolchain: ToolchainVersion, currentToolchain: ToolchainVersion?) throws -> Bool {
+        let toolchainBinURL = self.swiftlyToolchainsDir
             .appendingPathComponent(toolchain.name, isDirectory: true)
             .appendingPathComponent("usr", isDirectory: true)
             .appendingPathComponent("bin", isDirectory: true)
 
         // Delete existing symlinks from previously in-use toolchain.
-        for existingExecutable in try FileManager.default.contentsOfDirectory(atPath: SwiftlyCore.binDir.path) {
-            guard existingExecutable != "swiftly" else {
-                continue
-            }
-            try SwiftlyCore.binDir.appendingPathComponent(existingExecutable).deleteIfExists()
+        if let currentToolchain {
+            try self.unUse(currentToolchain: currentToolchain)
         }
 
-        for executable in try FileManager.default.contentsOfDirectory(atPath: toolchainBinURL.path) {
-            let linkURL = SwiftlyCore.binDir.appendingPathComponent(executable)
+        // Ensure swiftly doesn't overwrite any existing executables without getting confirmation first.
+        let swiftlyBinDirContents = try FileManager.default.contentsOfDirectory(atPath: self.swiftlyBinDir.path)
+        let toolchainBinDirContents = try FileManager.default.contentsOfDirectory(atPath: toolchainBinURL.path)
+        let willBeOverwritten = Set(toolchainBinDirContents).intersection(swiftlyBinDirContents)
+        if !willBeOverwritten.isEmpty {
+            SwiftlyCore.print("The following existing executables will be overwritten:")
+
+            for executable in willBeOverwritten {
+                SwiftlyCore.print("  \(self.swiftlyBinDir.appendingPathComponent(executable).path)")
+            }
+
+            let proceed = SwiftlyCore.readLine(prompt: "Proceed? (y/n)") ?? "n"
+
+            guard proceed == "y" else {
+                SwiftlyCore.print("Aborting use")
+                return false
+            }
+        }
+
+        for executable in toolchainBinDirContents {
+            let linkURL = self.swiftlyBinDir.appendingPathComponent(executable)
             let executableURL = toolchainBinURL.appendingPathComponent(executable)
 
+            // Deletion confirmed with user above.
             try linkURL.deleteIfExists()
 
             try FileManager.default.createSymbolicLink(
                 atPath: linkURL.path,
                 withDestinationPath: executableURL.path
             )
+        }
+
+        return true
+    }
+
+    public func unUse(currentToolchain: ToolchainVersion) throws {
+        let currentToolchainBinURL = self.swiftlyToolchainsDir
+            .appendingPathComponent(currentToolchain.name, isDirectory: true)
+            .appendingPathComponent("usr", isDirectory: true)
+            .appendingPathComponent("bin", isDirectory: true)
+
+        for existingExecutable in try FileManager.default.contentsOfDirectory(atPath: currentToolchainBinURL.path) {
+            guard existingExecutable != "swiftly" else {
+                continue
+            }
+
+            let url = self.swiftlyBinDir.appendingPathComponent(existingExecutable)
+            let vals = try url.resourceValues(forKeys: [.isSymbolicLinkKey])
+
+            guard let islink = vals.isSymbolicLink, islink else {
+                throw Error(message: "Found executable not managed by swiftly in SWIFTLY_BIN_DIR: \(url.path)")
+            }
+            let symlinkDest = url.resolvingSymlinksInPath()
+            guard symlinkDest.deletingLastPathComponent() == currentToolchainBinURL else {
+                throw Error(message: "Found symlink that points to non-swiftly managed executable: \(symlinkDest.path)")
+            }
+
+            try self.swiftlyBinDir.appendingPathComponent(existingExecutable).deleteIfExists()
         }
     }
 
@@ -101,12 +139,5 @@ public struct Linux: Platform {
         FileManager.default.temporaryDirectory.appendingPathComponent("swiftly-\(UUID())")
     }
 
-    public static let currentPlatform: any Platform = {
-        do {
-            let config = try Config.load()
-            return Linux(platform: config.platform)
-        } catch {
-            fatalError("error loading config: \(error)")
-        }
-    }()
+    public static let currentPlatform: any Platform = Linux()
 }
